+++ src/features/settings/SettingsNavigator.tsx
@@ -2,13 +2,16 @@ import React from 'react';
 
 import { RootStack } from 'src/navigation';
 
-import ProfileSettingsScreen from './ProfileSettingsScreen';
-import NotificationsSettingsScreen from './NotificationSettings';
 import MainSettingsScreen from './MainSettingsScreen';
+import ProfileSettingsScreen from './ProfileSettingsScreen';
+import AccountTypeSettingsScreen from './AccountTypeSettingsScreen';
+import NotificationsSettingsScreen from './NotificationSettings';
+import { color } from 'src/constants';
 
 export default function renderSettingsNavigator() {
   return (
-    <RootStack.Group>
+    <RootStack.Group
+      screenOptions={{ cardStyle: { backgroundColor: color.white } }}>
       <RootStack.Screen
         name="MainSettings"
         component={MainSettingsScreen}
@@ -18,6 +21,11 @@ export default function renderSettingsNavigator() {
         name="ProfileSettings"
         component={ProfileSettingsScreen}
         options={{ title: 'Profile Settings' }}
+      />
+      <RootStack.Screen
+        name="AccountTypeSettings"
+        component={AccountTypeSettingsScreen}
+        options={{ title: 'Account Type' }}
       />
       <RootStack.Screen
         name="NotificationSettings"